+++ nise/yaml_gen.py
@@ -1,21 +1,3 @@
-#! /usr/bin/env python3
-#
-# Copyright 2020 Red Hat, Inc.
-#
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU Affero General Public License as
-# published by the Free Software Foundation, either version 3 of the
-# License, or (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU Affero General Public License for more details.
-#
-# You should have received a copy of the GNU Affero General Public License
-# along with this program.  If not, see <https://www.gnu.org/licenses/>.
-#
-"""Utility to generate large yaml files."""
 import logging
 import os
 from argparse import ArgumentParser
@@ -23,6 +5,7 @@ from argparse import ArgumentParser
 from dateutil.parser import parse
 from nise.yaml_generator.aws.generator import AWSGenerator
 from nise.yaml_generator.ocp.generator import OCPGenerator
+
 
 GENERATOR_MAP = {"AWS": AWSGenerator(), "OCP": OCPGenerator()}
